executor/lockstats: reuse the same code and return err earlier
